Refactor navigation data to use WooPosVariableProductNavigationData across relevant components
