+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/items/navigation/WooPosItemsNavigator.kt
@@ -1,6 +1,6 @@
 package com.woocommerce.android.ui.woopos.home.items.navigation
 
-import com.woocommerce.android.ui.woopos.home.items.WooPosItemNavigationData.VariableProductData
+import com.woocommerce.android.ui.woopos.home.items.variations.WooPosVariableProductNavigationData
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import javax.inject.Inject
@@ -14,7 +14,7 @@ class WooPosItemsNavigator @Inject constructor() {
     }
 
     sealed class WooPosItemsScreenNavigationEvent {
-        data class NavigateToVariationsScreen(val product: VariableProductData) : WooPosItemsScreenNavigationEvent()
+        data class NavigateToVariationsScreen(val product: WooPosVariableProductNavigationData) : WooPosItemsScreenNavigationEvent()
         data object NavigateBackToItemListScreen : WooPosItemsScreenNavigationEvent()
     }
 }
